refactor(Program): update logging and service modules to use scoped Var::Type enum
- Update AppLog to use scoped Var::Type enum values for log variable handling
- Update RpcLog and WebLog to use scoped Var::Type constants
- Update Application service class to use scoped Var::Type for configuration
- Update GlobalArguments parser to use scoped Var::Type when processing
  command-line arguments and environment variables
- Ensures program-level variable handling is consistent with Memory module refactoring

+++ Mantids30/Program_Logs/applog.h
@@ -17,7 +17,7 @@ public:
      * Class constructor.
      * @param _logMode Log mode (or combination)
      */
-    AppLog(const uint8_t & _logMode = static_cast<unsigned int>(Mode::STANDARD));
+    AppLog(const uint8_t & _logMode = Mode::STANDARD);
 
     /**
      * Log an application event with user and IP information and max out size.
